old_configs/requirements.txt
absl-py==1.4.0
astunparse==1.6.3
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.8
fonttools==4.39.4
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.9
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mysql-connector-python
mysqlclient
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyodbc==5.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.15
tenacity==8.2.2
tensorboard-data-server==0.7.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.3
wrapt==1.14.1
